refactor: split the background text editor into its own module
